+++ frontend/src/app/app.config.ts
@@ -1,4 +1,4 @@
-import { ApplicationConfig } from '@angular/core';
+import { APP_INITIALIZER, ApplicationConfig } from '@angular/core';
 import { provideRouter } from '@angular/router';
 
 import { routes } from './app.routes';
@@ -8,6 +8,7 @@ import localeFRCH from '@angular/common/locales/fr-CH';
 import { registerLocaleData } from '@angular/common';
 import { MatPaginatorIntl } from '@angular/material/paginator';
 import { MatPaginatorIntlCevi } from './core/internationalization/MatPaginatorIntlCevi';
+import { ConfigService } from './core/services/config.service';
 
 registerLocaleData(localeDECH);
 registerLocaleData(localeFRCH);
@@ -17,5 +18,11 @@ export const appConfig: ApplicationConfig = {
     provideRouter(routes),
     provideHttpClient(),
     { provide: MatPaginatorIntl, useClass: MatPaginatorIntlCevi },
+    {
+      provide: APP_INITIALIZER,
+      useFactory: (cfg: ConfigService) => () => cfg.load(),
+      deps: [ConfigService],
+      multi: true,
+    },
   ],
 };