+++ frontend/src/environments/environment.int.ts
@@ -1,3 +1,0 @@
-export const environment = {
-  apiUri: 'https://events-api-int.cevi.tools',
-};