+++ frontend/src/environments/environment.prod.ts
@@ -1,3 +1,0 @@
-export const environment = {
-  apiUri: 'https://events-api.cevi.tools',
-};