+++ frontend/src/environments/environment.ts
@@ -1,4 +1,0 @@
-// local
-export const environment = {
-  apiUri: 'http://localhost:8080',
-};